Issue #14: LEADER handling duplicating the field
This patch makes the handling of the LEADER field use $record->leader to
set the LEADER instead of creating a new MARC::Field object to be
inserted into the fresh MARC::Record.

This seems to be the way to set the LEADER on new records, otherwise you
end up duplicating the 000 field.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>